requirements.txt
PySide6
Pillow
onnxruntime==1.17.1